feat(fase-07): close 7.4 lock-free readers + 7.4a production-safe storage
- Fix doctest: Mutex → RwLock in axiomdb-network lib.rs example
- Fix clippy: remove unnecessary u64 cast in pwrite_page
- Fix clippy: gate flush_runs under #[cfg(test)]
- Update storage.md: document pwrite write path, read-only mmap,
  owned PageRef, deferred free queue, fsync flush
- Update mvcc.md: document lock-free reader architecture,
  RwLock concurrency model, zero-lock read advantage callout,
  fix default isolation level to REPEATABLE READ
- Mark 7.4 as completed in progreso.md

Phase 7/34 in progress. See specs/fase-07/
Spec: specs/fase-07/spec-7.4*.md | Tests: 1828 passed

+++ crates/axiomdb-storage/src/mmap.rs
@@ -300,7 +300,7 @@ impl MmapStorage {
 
     /// Writes a full page via `pwrite()` at the correct file offset.
     fn pwrite_page(&self, page_id: u64, page: &Page) -> Result<(), DbError> {
-        let offset = page_id as u64 * PAGE_SIZE as u64;
+        let offset = page_id * PAGE_SIZE as u64;
         self.pwrite_bytes(offset, page.as_bytes())
     }
 
@@ -360,6 +360,7 @@ impl MmapStorage {
     ///
     /// Returns on the first error without calling the remaining entries.
     /// This makes the failure path testable: pass an injected flusher in tests.
+    #[cfg(test)]
     fn flush_runs<F>(runs: &[(usize, usize)], mut flusher: F) -> std::io::Result<()>
     where
         F: FnMut(usize, usize) -> std::io::Result<()>,
